Reformat commit 6b2a1445ec8a631060c4cbff3f172bf31d3379b9
This commit broke PDF builds on the 9.0 branch. Split the paragraph
up in two separate ones to make the PDFs happy again. Only doing this
in the 9.0 branch since that's the only place it's causing problems
at this point, and it's fairly unlikely that we're going to be
backpatching more docs changes to exactly this paragraph.